% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_df}
\alias{example_df}
\title{Example dataset.}
\format{
A data frame with 10 rows and 2 variables:
\describe{
  \item{x}{first variable, random sample, without replacement, of integers from 1 to 10}
  \item{y}{second variable, random sample of size 10 from standard normal distribution}
}
}
\source{
\code{example_df <- data.frame(x = sample(10), y = rnorm(10))}
}
\usage{
example_df
}
\description{
A dataset containing two randomly generated variables.
}
\keyword{datasets}
