% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_function.R
\name{toy_function}
\alias{toy_function}
\title{A toy function.}
\usage{
toy_function()
}
\value{
None. The function simply prints out a message.
}
\description{
This function prints out a simple, not very useful, message.
}
\examples{
toy_function()

}
\seealso{
The general \code{\link[base]{print}} function.
}
